requirements.txt
absl-py==1.4.0
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.2
blis==0.7.11
cachetools==5.3.0
catalogue==2.0.10
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.1.14
confection==0.1.4
contourpy==1.0.7
cryptography==41.0.7
cycler==0.11.0
cymem==2.0.8
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
dnspython==2.4.2
docx2txt==0.8
fastapi==0.95.2
filelock==3.12.0
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.9
fonttools==4.39.4
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.2
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.10
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.4
pdfid==1.1.3
pdfminer.six==20221105
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.5.0
preshed==3.0.9
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pyClamd==0.4.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.7
pymongo==4.6.1
pyparsing==3.0.9
pypiwin32==223
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pywin32==305
referencing==0.32.0
regex==2023.10.3
requests==2.30.0
requests-oauthlib==1.3.1
rpds-py==0.15.2
rsa==4.9
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.9.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.4.8
starlette==0.27.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
wasabi==1.1.2
weasel==0.3.4
Werkzeug==2.3.4
wrapt==1.14.1
